Fixed client connection errors when nodes have high weight and a crash when 'WSREP has not yet prepared node for application use'
If you use have nodes with high weights, you can have an overflow when calculating the
load in the nodes in a HG.

When a 'WSREP has not yet prepared node for application use' arises, proxysql deleted
the connection and tried to use it later

+++ include/MySQL_HostGroups_Manager.h
@@ -80,7 +80,7 @@ class MySrvC {	// MySQL Server Container
 	char *address;
 	uint16_t port;
 	uint16_t flags;
-	unsigned int weight;
+	uint64_t weight;
 	enum MySerStatus status;
 	unsigned int compression;
 	unsigned int max_connections;
